+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Tamr Terraform AWS Networking Repo
+## v1.1.0 - January 11st 2022
+* Adds Cloudwatch VPC Interface Endpoint.
+* Adds Cloudwatch VPC Interface Endpoint security group.
+* Adds a description to each output.
+
 ## v1.0.1 - December 30th 2021
 * Forces dependency between EMR and Networking module.